feat: integrate executable build pipeline and update core generator logic

+++ executable/build.bat
@@ -9,13 +9,14 @@ python -m pip install pyinstaller
 python -m pip install pywebview lxml xlrd openpyxl
 
 echo Removing old builds...
-rmdir /S /Q build
-rmdir /S /Q dist
+if exist build rmdir /S /Q build
+if exist dist rmdir /S /Q dist
 
 echo Initiating Compilation Sandbox...
 rem --noconsole hides the cmd window
 rem --onefile makes it a single executable payload
 python -m PyInstaller --noconfirm --clean --workpath "%TEMP%\cvsu_build" --distpath "%TEMP%\cvsu_dist" --name "CvSU Gen (Beta)" --onefile --windowed --icon "app_icon.ico" ^
+    --hidden-import pycparser.lextab --hidden-import pycparser.yacctab ^
     --paths ".." ^
     --add-data "ui.html;." ^
     --add-data "..\templates;templates/" ^
